+++ src/OpenTelemetryExporterTest.php
@@ -33,14 +33,14 @@ class OpenTelemetryExporterTest extends TestCase
      */
     private static function createSummary(string $label): Summary
     {
-        $start = new Snapshot(new DateTimeImmutable(), hrtime(true), [
+        $start = new Snapshot('start', new DateTimeImmutable(), hrtime(true), [
             'ru_utime.tv_sec' => 1,
             'ru_stime.tv_sec' => 1,
             'ru_utime.tv_usec' => 1,
             'ru_stime.tv_usec' => 1,
         ], 1000, 2000, 3000, 4000);
         usleep(100);
-        $end = new Snapshot(new DateTimeImmutable(), hrtime(true) + 1, [
+        $end = new Snapshot('end', new DateTimeImmutable(), hrtime(true) + 1, [
             'ru_utime.tv_sec' => 1,
             'ru_stime.tv_sec' => 1,
             'ru_utime.tv_usec' => 1,
@@ -117,7 +117,6 @@ class OpenTelemetryExporterTest extends TestCase
         $profiler = new Profiler(fn () => null);
         $reflection = new ReflectionClass($profiler);
         $reflection->getProperty('summaries')->setValue($profiler, [$summary1, $summary2]);
-        $reflection->getProperty('labels')->setValue($profiler, [$summary1->label => 1, $summary2->label => 1]);
 
         $this->exporter->exportProfiler($profiler);
         $spans = $this->otlExporter->getSpans();
@@ -150,21 +149,21 @@ class OpenTelemetryExporterTest extends TestCase
     public function it_can_export_a_marker(): void
     {
         $marker = new Marker('test-marker');
-        $start = new Snapshot(new DateTimeImmutable(), hrtime(true), [
+        $start = new Snapshot('start', new DateTimeImmutable(), hrtime(true), [
             'ru_utime.tv_sec' => 1,
             'ru_stime.tv_sec' => 1,
             'ru_utime.tv_usec' => 1,
             'ru_stime.tv_usec' => 1,
         ], 1000, 2000, 3000, 4000);
         usleep(100);
-        $middle = new Snapshot(new DateTimeImmutable(), hrtime(true), [
+        $middle = new Snapshot('middle', new DateTimeImmutable(), hrtime(true), [
             'ru_utime.tv_sec' => 1,
             'ru_stime.tv_sec' => 1,
             'ru_utime.tv_usec' => 1,
             'ru_stime.tv_usec' => 1,
         ], 1000, 2000, 3000, 4000);
         usleep(100);
-        $end = new Snapshot(new DateTimeImmutable(), hrtime(true) + 1, [
+        $end = new Snapshot('end', new DateTimeImmutable(), hrtime(true) + 1, [
             'ru_utime.tv_sec' => 1,
             'ru_stime.tv_sec' => 1,
             'ru_utime.tv_usec' => 1,